refactor

+++ src/features/users/api/users.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   Get,
   HttpCode,
+  HttpException,
   HttpStatus,
   Param,
   Post,
@@ -13,18 +14,19 @@ import { UsersService } from '../application/users.service';
 import { UsersQueryRepo } from '../infrastructure/users.query-repo';
 import { UsersInputBody } from './models/input/users.input.model';
 import { UsersQueryParamsInputModel } from './models/input/users-query-params.input.model';
-import { SortDirections } from '../../../common/types/interfaces';
 import { setPagination } from '../../../utils/set-pagination';
+import { UsersRepo } from '../infrastructure/users.repo';
 
 @Controller('users')
 export class UsersController {
   constructor(
     private usersService: UsersService,
-    private usersQueryRepo: UsersQueryRepo
+    private usersQueryRepo: UsersQueryRepo,
+    private usersRepo: UsersRepo
   ) {}
   @Get()
   @HttpCode(HttpStatus.OK)
-  getBlogs(
+  getUsers(
     @Query()
     queryParams: UsersQueryParamsInputModel
   ) {
@@ -40,6 +42,11 @@ export class UsersController {
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
   async deleteUser(@Param('id') id: string) {
+    const existsBlog = await this.usersRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Post with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     await this.usersService.deleteUser(id);
   }
 }
